Added the ability for home owners to expand their own houses for the cost of 1 gold token
git-svn-id: svn+ssh://4dimensions.org/home/game-source/repositories/gamesrc@501 254f5b4c-bdf4-4503-907e-934c7b59e499

+++ act.comm.c
@@ -1396,7 +1396,7 @@ ACMD(do_ctell) {
             return;
         }
         c = atoi (argument);
-        if ((c < 0) || (c >= clan.size())) {
+        if ((c < 0) || ((unsigned int)c >= clan.size())) {
             *ch << "There is no clan with that number.\r\n";
             return;
         } else {
